+++ src/remark-react/oneshot.tsx
@@ -4,22 +4,26 @@ import * as html from 'remark-html';
 import remarkReact from 'remark-react';
 
 export interface AppProps {
-  title: string;
+  body: string;
 }
 
 // @ts-ignore
 const processor = remark().use(remarkReact, {
   prefix: 'md-',
-  sanitize: true
-})
+  sanitize: true,
+});
 
 export class App extends React.Component<AppProps, {}> {
   public render() {
-    // @ts-ignore
-    const content = { __html: remark().use(remarkReact, {
-      sanitize: false
-    }).use(html).processSync(this.props.title).contents };
+    const content = {
+      // @ts-ignore
+      __html: remark()
+        .use(remarkReact, {
+          sanitize: false,
+        })
+        .use(html)
+        .processSync(this.props.body).contents,
+    };
     return <div dangerouslySetInnerHTML={content} />;
   }
 }
-